# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/k8s/service/create": {
            "post": {
                "description": "创建service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "创建service",
                "operationId": "/api/k8s/service/create",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kubeDto.ServiceCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"创建成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/service/del": {
            "delete": {
                "description": "删除service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "删除service",
                "operationId": "/api/k8s/service/del",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"删除成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/service/detail": {
            "get": {
                "description": "获取service详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "获取service详情",
                "operationId": "/api/k8s/service/detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\":v1.Deployment }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/service/list": {
            "get": {
                "description": "查看service列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "查看service列表",
                "operationId": "/api/k8s/service/list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤",
                        "name": "filter_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
        "/api/k8s/service/update": {
            "put": {
                "description": "更新service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "更新service",
                "operationId": "/api/k8s/service/update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "更新内容",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"更新成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
        "/api/k8s/workflow/create": {
            "post": {
                "description": "创建workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflow"
                ],
                "summary": "创建workflow",
                "operationId": "/api/k8s/workflow/create",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kubeDto.WorkFlowCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"创建成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/workflow/del": {
            "delete": {
                "description": "删除Workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflow"
                ],
                "summary": "删除Workflow",
                "operationId": "/api/k8s/workflow/del",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"删除成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/workflow/id": {
            "get": {
                "description": "根据ID查看workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflow"
                ],
                "summary": "根据ID查看workflow",
                "operationId": "/api/k8s/workflow/id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/workflow/list": {
            "get": {
                "description": "查看Configmap列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflow"
                ],
                "summary": "查看Configmap列表",
                "operationId": "/api/k8s/workflow/list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤",
                        "name": "filter_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kubeDto.HttpPath": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "path_type": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "service_port": {
                    "type": "integer"
                }
            }
        },
...
        "kubeDto.ServiceCreateInput": {
            "type": "object",
            "properties": {
                "container_port": {
                    "type": "integer"
                },
                "label": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "node_port": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "kubeDto.WorkFlowCreateInput": {
            "type": "object",
            "properties": {
                "container_port": {
                    "type": "integer"
                },
                "cpu": {
                    "type": "string"
                },
                "deployment": {
                    "type": "string"
                },
                "healthCheck": {
                    "type": "boolean"
                },
                "health_path": {
                    "type": "string"
                },
                "hosts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/kubeDto.HttpPath"
                        }
                    }
                },
                "image": {
                    "type": "string"
                },
                "label": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "memory": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "node_port": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "replicas": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "real_err": {
                    "type": "string"
                }
            }
        },
...
    }
}